@@ -2,18 +2,23 @@
 buildscript {
     repositories {
         mavenLocal()
+        mavenCentral()
+        google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.2.0'
+        classpath 'com.android.tools.build:gradle:3.2.1'
         classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.6'
-        classpath "com.github.dcendents:android-maven-gradle-plugin:1.4.1"
+        classpath "com.github.dcendents:android-maven-gradle-plugin:2.1"
+        classpath "com.github.jengelman.gradle.plugins:shadow:4.0.4"
     }
 }
 
 allprojects {
     repositories {
         mavenLocal()
+        mavenCentral()
+        google()
         jcenter()
     }
 
@@ -25,10 +30,10 @@ allprojects {
 
 ext {
     minSdkVersion = 10
-    compileSdkVersion = 23
-    targetSdkVersion = compileSdkVersion
-    buildToolsVersion = '23.0.2'
-    supportLibVersion = '23.0.1'
+    compileSdkVersion = 29
+    targetSdkVersion = 23
+    buildToolsVersion = '28.0.3'
+    androidXAnnotationVersion = '1.1.0'
     javaVersion = JavaVersion.VERSION_1_7
 
     GROUP = 'com.tencent.tinker'
@@ -52,3 +57,5 @@ ext {
     BINTRAY_ORGANIZATION = "tinker"
 
 }
+
+apply from: rootProject.file('gradle/check.gradle')
