@@ -17,6 +17,8 @@
 package com.tencent.tinker.android.dex;
 
 import com.tencent.tinker.android.dex.TableOfContents.Section.Item;
+import com.tencent.tinker.android.dex.util.CompareUtils;
+import com.tencent.tinker.android.dex.util.HashCodeHelper;
 
 import static com.tencent.tinker.android.dex.EncodedValueReader.ENCODED_ANNOTATION;
 
@@ -44,7 +46,22 @@ public final class Annotation extends Item<Annotation> {
     }
 
     @Override public int compareTo(Annotation other) {
-        return encodedAnnotation.compareTo(other.encodedAnnotation);
+        int cmpRes = encodedAnnotation.compareTo(other.encodedAnnotation);
+        if (cmpRes != 0) return cmpRes;
+        return CompareUtils.uCompare(visibility, other.visibility);
+    }
+
+    @Override
+    public int hashCode() {
+        return HashCodeHelper.hash(visibility, encodedAnnotation);
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (!(obj instanceof Annotation)) {
+            return false;
+        }
+        return this.compareTo((Annotation) obj) == 0;
     }
 
     @Override
