@@ -16,7 +16,7 @@
 
 package com.tencent.tinker.lib.patch;
 
-import com.tencent.tinker.commons.util.StreamUtil;
+import com.tencent.tinker.commons.util.IOHelper;
 import com.tencent.tinker.lib.util.TinkerLog;
 import com.tencent.tinker.loader.shareutil.ShareConstants;
 import com.tencent.tinker.loader.shareutil.SharePatchFileUtil;
@@ -44,11 +44,13 @@ public class BasePatchInternal {
     protected static final String DEX_META_FILE        = ShareConstants.DEX_META_FILE;
     protected static final String SO_META_FILE         = ShareConstants.SO_META_FILE;
     protected static final String RES_META_FILE        = ShareConstants.RES_META_FILE;
+    protected static final String ARKHOT_META_FILE = ShareConstants.ARKHOT_META_FILE;
 
     protected static final int TYPE_DEX         = ShareConstants.TYPE_DEX;
-    protected static final int TYPE_Library     = ShareConstants.TYPE_LIBRARY;
+    protected static final int TYPE_LIBRARY     = ShareConstants.TYPE_LIBRARY;
     protected static final int TYPE_RESOURCE    = ShareConstants.TYPE_RESOURCE;
     protected static final int TYPE_CLASS_N_DEX = ShareConstants.TYPE_CLASS_N_DEX;
+    protected static final int TYPE_ARKHOT_SO = ShareConstants.TYPE_ARKHOT_SO;
 
 
     public static boolean extract(ZipFile zipFile, ZipEntry entryFile, File extractTo, String targetMd5, boolean isDex) throws IOException {
@@ -70,8 +72,8 @@ public class BasePatchInternal {
                     os.write(buffer, 0, length);
                 }
             } finally {
-                StreamUtil.closeQuietly(os);
-                StreamUtil.closeQuietly(is);
+                IOHelper.closeQuietly(os);
+                IOHelper.closeQuietly(is);
             }
             if (targetMd5 != null) {
                 if (isDex) {
@@ -99,7 +101,7 @@ public class BasePatchInternal {
     public static int getMetaCorruptedCode(int type) {
         if (type == TYPE_DEX) {
             return ShareConstants.ERROR_PACKAGE_CHECK_DEX_META_CORRUPTED;
-        } else if (type == TYPE_Library) {
+        } else if (type == TYPE_LIBRARY) {
             return ShareConstants.ERROR_PACKAGE_CHECK_LIB_META_CORRUPTED;
         } else if (type == TYPE_RESOURCE) {
             return ShareConstants.ERROR_PACKAGE_CHECK_RESOURCE_META_CORRUPTED;
