@@ -17,7 +17,6 @@
 
 package com.tencent.tinker.loader;
 
-import android.annotation.SuppressLint;
 import android.app.Application;
 import android.os.Build;
 import android.util.Log;
@@ -52,27 +51,27 @@ public class SystemClassLoaderAdder {
     private static final String TAG = "Tinker.ClassLoaderAdder";
     private static int sPatchDexCount = 0;
 
-    @SuppressLint("NewApi")
-    public static void installDexes(Application application, PathClassLoader loader, File dexOptDir, List<File> files)
+    public static void installDexes(Application application, ClassLoader loader, File dexOptDir, List<File> files, boolean isProtectedApp)
         throws Throwable {
         Log.i(TAG, "installDexes dexOptDir: " + dexOptDir.getAbsolutePath() + ", dex size:" + files.size());
 
         if (!files.isEmpty()) {
             files = createSortedAdditionalPathEntries(files);
             ClassLoader classLoader = loader;
-            if (Build.VERSION.SDK_INT >= 24 && !checkIsProtectedApp(files)) {
-                classLoader = AndroidNClassLoader.inject(loader, application);
-            }
-            //because in dalvik, if inner class is not the same classloader with it wrapper class.
-            //it won't fail at dex2opt
-            if (Build.VERSION.SDK_INT >= 23) {
-                V23.install(classLoader, files, dexOptDir);
-            } else if (Build.VERSION.SDK_INT >= 19) {
-                V19.install(classLoader, files, dexOptDir);
-            } else if (Build.VERSION.SDK_INT >= 14) {
-                V14.install(classLoader, files, dexOptDir);
+            if (Build.VERSION.SDK_INT >= 24 && !isProtectedApp) {
+                classLoader = NewClassLoaderInjector.inject(application, loader, dexOptDir, files);
             } else {
-                V4.install(classLoader, files, dexOptDir);
+                //because in dalvik, if inner class is not the same classloader with it wrapper class.
+                //it won't fail at dex2opt
+                if (Build.VERSION.SDK_INT >= 23) {
+                    V23.install(classLoader, files, dexOptDir);
+                } else if (Build.VERSION.SDK_INT >= 19) {
+                    V19.install(classLoader, files, dexOptDir);
+                } else if (Build.VERSION.SDK_INT >= 14) {
+                    V14.install(classLoader, files, dexOptDir);
+                } else {
+                    V4.install(classLoader, files, dexOptDir);
+                }
             }
             //install done
             sPatchDexCount = files.size();
@@ -82,6 +81,22 @@ public class SystemClassLoaderAdder {
                 //reset patch dex
                 SystemClassLoaderAdder.uninstallPatchDex(classLoader);
                 throw new TinkerRuntimeException(ShareConstants.CHECK_DEX_INSTALL_FAIL);
+            }
+        }
+    }
+
+    public static void installApk(PathClassLoader loader, List<File> files) throws Throwable {
+        if (!files.isEmpty()) {
+            files = createSortedAdditionalPathEntries(files);
+            ClassLoader classLoader = loader;
+            ArkHot.install(classLoader, files);
+            sPatchDexCount = files.size();
+            Log.i(TAG, "after loaded classloader: " + classLoader + ", dex size:" + sPatchDexCount);
+
+            if (!checkDexInstall(classLoader)) {
+                // reset patch dex
+//                SystemClassLoaderAdder.uninstallPatchDex(classLoader);
+//                throw new TinkerRuntimeException(ShareConstants.CHECK_DEX_INSTALL_FAIL);
             }
         }
     }
@@ -101,6 +116,7 @@ public class SystemClassLoaderAdder {
             try {
                 ShareReflectUtil.reduceFieldArray(classLoader, "mDexs", sPatchDexCount);
             } catch (Exception e) {
+                // Ignored.
             }
         }
     }
@@ -111,20 +127,6 @@ public class SystemClassLoaderAdder {
         boolean isPatch = (boolean) filed.get(null);
         Log.w(TAG, "checkDexInstall result:" + isPatch);
         return isPatch;
-    }
-
-    private static boolean checkIsProtectedApp(List<File> files) {
-        if (!files.isEmpty()) {
-            for (File file : files) {
-                if (file == null) {
-                    continue;
-                }
-                if (file.getName().startsWith(ShareConstants.CHANGED_CLASSES_DEX_NAME)) {
-                    return true;
-                }
-            }
-        }
-        return false;
     }
 
     private static List<File> createSortedAdditionalPathEntries(List<File> additionalPathEntries) {
@@ -182,6 +184,27 @@ public class SystemClassLoaderAdder {
         });
 
         return result;
+    }
+
+    /**
+     * Installer for platform huawei ark
+     */
+    private static final class ArkHot {
+        private static void install(ClassLoader loader, List<File> additionalClassPathEntries)
+                throws IllegalArgumentException, IllegalAccessException, NoSuchMethodException,
+                InvocationTargetException, IOException, ClassNotFoundException, SecurityException {
+            Class<?>  extendedClassLoaderHelper = ClassLoader.getSystemClassLoader()
+                    .getParent().loadClass("com.huawei.ark.classloader.ExtendedClassLoaderHelper");
+
+            for (File file : additionalClassPathEntries) {
+                String path = file.getCanonicalPath();
+                Method applyPatchMethod = extendedClassLoaderHelper.getDeclaredMethod(
+                        "applyPatch", ClassLoader.class, String.class);
+                applyPatchMethod.setAccessible(true);
+                applyPatchMethod.invoke(null, loader, path);
+                Log.i(TAG, "ArkHot install path = " + path);
+            }
+        }
     }
 
     /**
@@ -379,7 +402,7 @@ public class SystemClassLoaderAdder {
             try {
                 ShareReflectUtil.expandFieldArray(loader, "mDexs", extraDexs);
             } catch (Exception e) {
-
+                // Ignored.
             }
         }
     }
