@@ -16,9 +16,7 @@
 
 package com.tencent.tinker.loader;
 
-import android.annotation.TargetApi;
 import android.content.Intent;
-import android.os.Build;
 import android.util.Log;
 
 import com.tencent.tinker.loader.app.TinkerApplication;
@@ -34,8 +32,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 
-import dalvik.system.PathClassLoader;
-
 /**
  * Created by zhangshaowen on 16/3/8.
  * check the complete of the dex files
@@ -50,10 +46,10 @@ public class TinkerDexLoader {
     private static final String DEFAULT_DEX_OPTIMIZE_PATH   = ShareConstants.DEFAULT_DEX_OPTIMIZE_PATH;
     private static final String INTERPRET_DEX_OPTIMIZE_PATH = ShareConstants.INTERPRET_DEX_OPTIMIZE_PATH;
 
-    private static final ArrayList<ShareDexDiffPatchInfo> loadDexList = new ArrayList<>();
+    private static final ArrayList<ShareDexDiffPatchInfo> LOAD_DEX_LIST = new ArrayList<>();
 
 
-    //    private static File testOptDexFile;
+    // private static File testOptDexFile;
     private static HashSet<ShareDexDiffPatchInfo> classNDexInfo = new HashSet<>();
 
     private static boolean isVmArt = ShareTinkerInternals.isVmArt();
@@ -67,14 +63,13 @@ public class TinkerDexLoader {
      *
      * @param application The application.
      */
-    @TargetApi(Build.VERSION_CODES.ICE_CREAM_SANDWICH)
-    public static boolean loadTinkerJars(final TinkerApplication application, String directory, String oatDir, Intent intentResult, boolean isSystemOTA) {
-        if (loadDexList.isEmpty() && classNDexInfo.isEmpty()) {
+    public static boolean loadTinkerJars(final TinkerApplication application, String directory, String oatDir, Intent intentResult, boolean isSystemOTA, boolean isProtectedApp) {
+        if (LOAD_DEX_LIST.isEmpty() && classNDexInfo.isEmpty()) {
             Log.w(TAG, "there is no dex to load");
             return true;
         }
 
-        PathClassLoader classLoader = (PathClassLoader) TinkerDexLoader.class.getClassLoader();
+        ClassLoader classLoader = TinkerDexLoader.class.getClassLoader();
         if (classLoader != null) {
             Log.i(TAG, "classloader: " + classLoader.toString());
         } else {
@@ -86,7 +81,7 @@ public class TinkerDexLoader {
 
         ArrayList<File> legalFiles = new ArrayList<>();
 
-        for (ShareDexDiffPatchInfo info : loadDexList) {
+        for (ShareDexDiffPatchInfo info : LOAD_DEX_LIST) {
             //for dalvik, ignore art support dex
             if (isJustArtSupportDex(info)) {
                 continue;
@@ -138,16 +133,16 @@ public class TinkerDexLoader {
                 targetISA = ShareTinkerInternals.getCurrentInstructionSet();
             } catch (Throwable throwable) {
                 Log.i(TAG, "getCurrentInstructionSet fail:" + throwable);
-//                try {
-//                    targetISA = ShareOatUtil.getOatFileInstructionSet(testOptDexFile);
-//                } catch (Throwable throwable) {
+                // try {
+                //     targetISA = ShareOatUtil.getOatFileInstructionSet(testOptDexFile);
+                // } catch (Throwable throwable) {
                 // don't ota on the front
                 deleteOutOfDateOATFile(directory);
 
                 intentResult.putExtra(ShareIntentUtil.INTENT_PATCH_INTERPRET_EXCEPTION, throwable);
                 ShareIntentUtil.setIntentReturnCode(intentResult, ShareConstants.ERROR_LOAD_PATCH_GET_OTA_INSTRUCTION_SET_EXCEPTION);
                 return false;
-//                }
+                // }
             }
 
             deleteOutOfDateOATFile(directory);
@@ -157,7 +152,7 @@ public class TinkerDexLoader {
             optimizeDir = new File(directory + "/" + INTERPRET_DEX_OPTIMIZE_PATH);
 
             TinkerDexOptimizer.optimizeAll(
-                legalFiles, optimizeDir, true, targetISA,
+                application, legalFiles, optimizeDir, true, targetISA,
                 new TinkerDexOptimizer.ResultCallback() {
                     long start;
 
@@ -191,10 +186,9 @@ public class TinkerDexLoader {
             }
         }
         try {
-            SystemClassLoaderAdder.installDexes(application, classLoader, optimizeDir, legalFiles);
+            SystemClassLoaderAdder.installDexes(application, classLoader, optimizeDir, legalFiles, isProtectedApp);
         } catch (Throwable e) {
             Log.e(TAG, "install dexes failed");
-//            e.printStackTrace();
             intentResult.putExtra(ShareIntentUtil.INTENT_PATCH_EXCEPTION, e);
             ShareIntentUtil.setIntentReturnCode(intentResult, ShareConstants.ERROR_LOAD_PATCH_VERSION_DEX_LOAD_EXCEPTION);
             return false;
@@ -215,7 +209,7 @@ public class TinkerDexLoader {
         if (meta == null) {
             return true;
         }
-        loadDexList.clear();
+        LOAD_DEX_LIST.clear();
         classNDexInfo.clear();
 
         ArrayList<ShareDexDiffPatchInfo> allDexInfo = new ArrayList<>();
@@ -245,7 +239,7 @@ public class TinkerDexLoader {
                 classNDexInfo.add(info);
             } else {
                 dexes.put(info.realName, getInfoMd5(info));
-                loadDexList.add(info);
+                LOAD_DEX_LIST.add(info);
             }
         }
 
@@ -280,14 +274,17 @@ public class TinkerDexLoader {
             //check dex opt whether complete also
             File dexOptFile = new File(SharePatchFileUtil.optimizedPathFor(dexFile, optimizeDexDirectoryFile));
             if (!SharePatchFileUtil.isLegalFile(dexOptFile)) {
+                if (SharePatchFileUtil.shouldAcceptEvenIfIllegal(dexOptFile)) {
+                    continue;
+                }
                 intentResult.putExtra(ShareIntentUtil.INTENT_PATCH_MISSING_DEX_PATH, dexOptFile.getAbsolutePath());
                 ShareIntentUtil.setIntentReturnCode(intentResult, ShareConstants.ERROR_LOAD_PATCH_VERSION_DEX_OPT_FILE_NOT_EXIST);
                 return false;
             }
-//            // find test dex
-//            if (dexOptFile.getName().startsWith(ShareConstants.TEST_DEX_NAME)) {
-//                testOptDexFile = dexOptFile;
-//            }
+            // // find test dex
+            // if (dexOptFile.getName().startsWith(ShareConstants.TEST_DEX_NAME)) {
+            //     testOptDexFile = dexOptFile;
+            // }
         }
 
         //if is ok, add to result intent
