@@ -16,9 +16,7 @@
 
 package com.tencent.tinker.loader;
 
-import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Build;
 import android.os.SystemClock;
 import android.util.Log;
@@ -52,6 +50,7 @@ public class TinkerLoader extends AbstractTinkerLoader {
      */
     @Override
     public Intent tryLoad(TinkerApplication app) {
+        Log.d(TAG, "tryLoad test test");
         Intent resultIntent = new Intent();
 
         long begin = SystemClock.elapsedRealtime();
@@ -73,7 +72,6 @@ public class TinkerLoader extends AbstractTinkerLoader {
             Log.w(TAG, "tryLoadPatchFiles: we don't load patch with :patch process itself, just return");
             ShareIntentUtil.setIntentReturnCode(resultIntent, ShareConstants.ERROR_LOAD_DISABLE);
             return;
-
         }
         //tinker
         File patchDirectoryFile = SharePatchFileUtil.getPatchDirectory(app);
@@ -111,6 +109,9 @@ public class TinkerLoader extends AbstractTinkerLoader {
             return;
         }
 
+        final boolean isProtectedApp = patchInfo.isProtectedApp;
+        resultIntent.putExtra(ShareIntentUtil.INTENT_IS_PROTECTED_APP, isProtectedApp);
+
         String oldVersion = patchInfo.oldVersion;
         String newVersion = patchInfo.newVersion;
         String oatDex = patchInfo.oatDir;
@@ -122,17 +123,62 @@ public class TinkerLoader extends AbstractTinkerLoader {
             return;
         }
 
+        boolean mainProcess = ShareTinkerInternals.isInMainProcess(app);
+        boolean isRemoveNewVersion = patchInfo.isRemoveNewVersion;
+
+        if (mainProcess) {
+            final String patchName = SharePatchFileUtil.getPatchVersionDirectory(newVersion);
+            // So far new version is not loaded in main process and other processes.
+            // We can remove new version directory safely.
+            if (isRemoveNewVersion) {
+                Log.w(TAG, "found clean patch mark and we are in main process, delete patch file now.");
+                if (patchName != null) {
+                    // oldVersion.equals(newVersion) means the new version has been loaded at least once
+                    // after it was applied.
+                    final boolean isNewVersionLoadedBefore = oldVersion.equals(newVersion);
+                    if (isNewVersionLoadedBefore) {
+                        // Set oldVersion and newVersion to empty string to clean patch
+                        // if current patch has been loaded before.
+                        oldVersion = "";
+                    }
+                    newVersion = oldVersion;
+                    patchInfo.oldVersion = oldVersion;
+                    patchInfo.newVersion = newVersion;
+                    patchInfo.isRemoveNewVersion = false;
+                    SharePatchInfo.rewritePatchInfoFileWithLock(patchInfoFile, patchInfo, patchInfoLockFile);
+
+                    String patchVersionDirFullPath = patchDirectoryPath + "/" + patchName;
+                    SharePatchFileUtil.deleteDir(patchVersionDirFullPath);
+
+                    if (isNewVersionLoadedBefore) {
+                        ShareTinkerInternals.killProcessExceptMain(app);
+                        ShareIntentUtil.setIntentReturnCode(resultIntent, ShareConstants.ERROR_LOAD_PATCH_DIRECTORY_NOT_EXIST);
+                        return;
+                    }
+                }
+            }
+            if (patchInfo.isRemoveInterpretOATDir) {
+                // delete interpret odex
+                // for android o, directory change. Fortunately, we don't need to support android o interpret mode any more
+                Log.i(TAG, "tryLoadPatchFiles: isRemoveInterpretOATDir is true, try to delete interpret optimize files");
+
+                patchInfo.isRemoveInterpretOATDir = false;
+                SharePatchInfo.rewritePatchInfoFileWithLock(patchInfoFile, patchInfo, patchInfoLockFile);
+                ShareTinkerInternals.killProcessExceptMain(app);
+                String patchVersionDirFullPath = patchDirectoryPath + "/" + patchName;
+                SharePatchFileUtil.deleteDir(patchVersionDirFullPath + "/" + ShareConstants.INTERPRET_DEX_OPTIMIZE_PATH);
+            }
+        }
+
         resultIntent.putExtra(ShareIntentUtil.INTENT_PATCH_OLD_VERSION, oldVersion);
         resultIntent.putExtra(ShareIntentUtil.INTENT_PATCH_NEW_VERSION, newVersion);
 
-        boolean mainProcess = ShareTinkerInternals.isInMainProcess(app);
         boolean versionChanged = !(oldVersion.equals(newVersion));
-        boolean oatModeChanged = oatDex.equals(ShareConstants.CHANING_DEX_OPTIMIZE_PATH) && mainProcess;
+        boolean oatModeChanged = oatDex.equals(ShareConstants.CHANING_DEX_OPTIMIZE_PATH);
         oatDex = ShareTinkerInternals.getCurrentOatMode(app, oatDex);
         resultIntent.putExtra(ShareIntentUtil.INTENT_PATCH_OAT_DIR, oatDex);
 
         String version = oldVersion;
-
         if (versionChanged && mainProcess) {
             version = newVersion;
         }
@@ -186,8 +232,9 @@ public class TinkerLoader extends AbstractTinkerLoader {
         resultIntent.putExtra(ShareIntentUtil.INTENT_PATCH_PACKAGE_CONFIG, securityCheck.getPackagePropertiesIfPresent());
 
         final boolean isEnabledForDex = ShareTinkerInternals.isTinkerEnabledForDex(tinkerFlag);
+        final boolean isArkHotRuning = ShareTinkerInternals.isArkHotRuning();
 
-        if (isEnabledForDex) {
+        if (!isArkHotRuning && isEnabledForDex) {
             //tinker/patch.info/patch-641e634c/dex
             boolean dexCheck = TinkerDexLoader.checkComplete(patchVersionDirectory, securityCheck, oatDex, resultIntent);
             if (!dexCheck) {
@@ -196,6 +243,17 @@ public class TinkerLoader extends AbstractTinkerLoader {
                 return;
             }
         }
+
+        final boolean isEnabledForArkHot = ShareTinkerInternals.isTinkerEnabledForArkHot(tinkerFlag);
+        if (isArkHotRuning && isEnabledForArkHot) {
+            boolean arkHotCheck = TinkerArkHotLoader.checkComplete(patchVersionDirectory, securityCheck, resultIntent);
+            if (!arkHotCheck) {
+                // file not found, do not load patch
+                Log.w(TAG, "tryLoadPatchFiles:dex check fail");
+                return;
+            }
+        }
+
 
         final boolean isEnabledForNativeLib = ShareTinkerInternals.isTinkerEnabledForNativeLib(tinkerFlag);
 
@@ -229,24 +287,16 @@ public class TinkerLoader extends AbstractTinkerLoader {
         resultIntent.putExtra(ShareIntentUtil.INTENT_PATCH_SYSTEM_OTA, isSystemOTA);
 
         //we should first try rewrite patch info file, if there is a error, we can't load jar
-        if ((mainProcess && versionChanged)
-             || oatModeChanged) {
-            patchInfo.oldVersion = version;
-            patchInfo.oatDir = oatDex;
-
-            //update old version to new
-            if (!SharePatchInfo.rewritePatchInfoFileWithLock(patchInfoFile, patchInfo, patchInfoLockFile)) {
-                ShareIntentUtil.setIntentReturnCode(resultIntent, ShareConstants.ERROR_LOAD_PATCH_REWRITE_PATCH_INFO_FAIL);
-                Log.w(TAG, "tryLoadPatchFiles:onReWritePatchInfoCorrupted");
-                return;
+        if (mainProcess) {
+            if (versionChanged) {
+                patchInfo.oldVersion = version;
             }
             if (oatModeChanged) {
-                // delete interpret odex
-                // for android o, directory change. Fortunately, we don't need to support android o interpret mode any more
-                Log.i(TAG, "tryLoadPatchFiles:oatModeChanged, try to delete interpret optimize files");
-                SharePatchFileUtil.deleteDir(patchVersionDirectory + "/" + ShareConstants.INTERPRET_DEX_OPTIMIZE_PATH);
+                patchInfo.oatDir = oatDex;
+                patchInfo.isRemoveInterpretOATDir = true;
             }
         }
+
         if (!checkSafeModeCount(app)) {
             resultIntent.putExtra(ShareIntentUtil.INTENT_PATCH_EXCEPTION, new TinkerRuntimeException("checkSafeModeCount fail"));
             ShareIntentUtil.setIntentReturnCode(resultIntent, ShareConstants.ERROR_LOAD_PATCH_UNCAUGHT_EXCEPTION);
@@ -255,8 +305,8 @@ public class TinkerLoader extends AbstractTinkerLoader {
         }
 
         //now we can load patch jar
-        if (isEnabledForDex) {
-            boolean loadTinkerJars = TinkerDexLoader.loadTinkerJars(app, patchVersionDirectory, oatDex, resultIntent, isSystemOTA);
+        if (!isArkHotRuning && isEnabledForDex) {
+            boolean loadTinkerJars = TinkerDexLoader.loadTinkerJars(app, patchVersionDirectory, oatDex, resultIntent, isSystemOTA, isProtectedApp);
 
             if (isSystemOTA) {
                 // update fingerprint after load success
@@ -279,6 +329,14 @@ public class TinkerLoader extends AbstractTinkerLoader {
             }
         }
 
+        if (isArkHotRuning && isEnabledForArkHot) {
+            boolean loadArkHotFixJars = TinkerArkHotLoader.loadTinkerArkHot(app, patchVersionDirectory, resultIntent);
+            if (!loadArkHotFixJars) {
+                Log.w(TAG, "tryLoadPatchFiles:onPatchLoadArkApkFail");
+                return;
+            }
+        }
+
         //now we can load patch resource
         if (isEnabledForResource) {
             boolean loadTinkerResources = TinkerResourceLoader.loadTinkerResources(app, patchVersionDirectory, resultIntent);
@@ -289,34 +347,42 @@ public class TinkerLoader extends AbstractTinkerLoader {
         }
 
         // Init component hotplug support.
-        if (isEnabledForDex && isEnabledForResource) {
+        if ((isEnabledForDex || isEnabledForArkHot) && isEnabledForResource) {
             ComponentHotplug.install(app, securityCheck);
         }
 
-        // kill all other process if oat mode change
-        if (oatModeChanged) {
-            ShareTinkerInternals.killAllOtherProcess(app);
-            Log.i(TAG, "tryLoadPatchFiles:oatModeChanged, try to kill all other process");
+        if (!AppInfoChangedBlocker.tryStart(app)) {
+            Log.w(TAG, "tryLoadPatchFiles:AppInfoChangedBlocker install fail.");
+            ShareIntentUtil.setIntentReturnCode(resultIntent, ShareConstants.ERROR_LOAD_PATCH_BAIL_HACK_FAILURE);
+            return;
         }
+
+        // Before successfully exit, we should update stored version info and kill other process
+        // to make them load latest patch when we first applied newer one.
+        if (mainProcess && (versionChanged || oatModeChanged)) {
+            //update old version to new
+            if (!SharePatchInfo.rewritePatchInfoFileWithLock(patchInfoFile, patchInfo, patchInfoLockFile)) {
+                ShareIntentUtil.setIntentReturnCode(resultIntent, ShareConstants.ERROR_LOAD_PATCH_REWRITE_PATCH_INFO_FAIL);
+                Log.w(TAG, "tryLoadPatchFiles:onReWritePatchInfoCorrupted");
+                return;
+            }
+
+            ShareTinkerInternals.killProcessExceptMain(app);
+        }
+
         //all is ok!
         ShareIntentUtil.setIntentReturnCode(resultIntent, ShareConstants.ERROR_LOAD_OK);
         Log.i(TAG, "tryLoadPatchFiles: load end, ok!");
-        return;
     }
 
     private boolean checkSafeModeCount(TinkerApplication application) {
-        String processName = ShareTinkerInternals.getProcessName(application);
-        String preferName = ShareConstants.TINKER_OWN_PREFERENCE_CONFIG + processName;
-        //each process have its own SharedPreferences file
-        SharedPreferences sp = application.getSharedPreferences(preferName, Context.MODE_PRIVATE);
-        int count = sp.getInt(ShareConstants.TINKER_SAFE_MODE_COUNT, 0) + 1;
-        Log.w(TAG, "tinker safe mode preferName:" + preferName + " count:" + count);
-        if (count >= ShareConstants.TINKER_SAFE_MODE_MAX_COUNT) {
-            sp.edit().putInt(ShareConstants.TINKER_SAFE_MODE_COUNT, 0).commit();
+        int count = ShareTinkerInternals.getSafeModeCount(application);
+        if (count >= ShareConstants.TINKER_SAFE_MODE_MAX_COUNT - 1) {
+            ShareTinkerInternals.setSafeModeCount(application, 0);
             return false;
         }
         application.setUseSafeMode(true);
-        sp.edit().putInt(ShareConstants.TINKER_SAFE_MODE_COUNT, count).commit();
+        ShareTinkerInternals.setSafeModeCount(application, count + 1);
         return true;
     }
 }
