@@ -57,6 +57,7 @@ class TinkerResourcePatcher {
 
     // method
     private static Method addAssetPathMethod = null;
+    private static Method addAssetPathAsSharedLibraryMethod = null;
     private static Method ensureStringBlocksMethod = null;
 
     // field
@@ -95,11 +96,19 @@ class TinkerResourcePatcher {
         // Create a new AssetManager instance and point it to the resources
         final AssetManager assets = context.getAssets();
         addAssetPathMethod = findMethod(assets, "addAssetPath", String.class);
+        if (shouldAddSharedLibraryAssets(context.getApplicationInfo())) {
+            addAssetPathAsSharedLibraryMethod =
+                    findMethod(assets, "addAssetPathAsSharedLibrary", String.class);
+        }
 
         // Kitkat needs this method call, Lollipop doesn't. However, it doesn't seem to cause any harm
         // in L, so we do it unconditionally.
-        stringBlocksField = findField(assets, "mStringBlocks");
-        ensureStringBlocksMethod = findMethod(assets, "ensureStringBlocks");
+        try {
+            stringBlocksField = findField(assets, "mStringBlocks");
+            ensureStringBlocksMethod = findMethod(assets, "ensureStringBlocks");
+        } catch (Throwable ignored) {
+            // Ignored.
+        }
 
         // Use class fetched from instance to avoid some ROMs that use customized AssetManager
         // class. (e.g. Baidu OS)
@@ -195,10 +204,26 @@ class TinkerResourcePatcher {
             throw new IllegalStateException("Could not create new AssetManager");
         }
 
+        // Add SharedLibraries to AssetManager for resolve system resources not found issue
+        // This influence SharedLibrary Package ID
+        if (shouldAddSharedLibraryAssets(appInfo)) {
+            for (String sharedLibrary : appInfo.sharedLibraryFiles) {
+                if (!sharedLibrary.endsWith(".apk")) {
+                    continue;
+                }
+                if (((Integer) addAssetPathAsSharedLibraryMethod.invoke(newAssetManager, sharedLibrary)) == 0) {
+                    throw new IllegalStateException("AssetManager add SharedLibrary Fail");
+                }
+                Log.i(TAG, "addAssetPathAsSharedLibrary " + sharedLibrary);
+            }
+        }
+
         // Kitkat needs this method call, Lollipop doesn't. However, it doesn't seem to cause any harm
         // in L, so we do it unconditionally.
-        stringBlocksField.set(newAssetManager, null);
-        ensureStringBlocksMethod.invoke(newAssetManager);
+        if (stringBlocksField != null && ensureStringBlocksMethod != null) {
+            stringBlocksField.set(newAssetManager, null);
+            ensureStringBlocksMethod.invoke(newAssetManager);
+        }
 
         for (WeakReference<Resources> wr : references) {
             final Resources resources = wr.get();
@@ -232,6 +257,7 @@ class TinkerResourcePatcher {
                     publicSourceDirField.set(context.getApplicationInfo(), externalResourceFile);
                 }
             } catch (Throwable ignore) {
+                // Ignored.
             }
         }
 
@@ -279,5 +305,10 @@ class TinkerResourcePatcher {
         }
         Log.i(TAG, "checkResUpdate success, found test resource assets file " + TEST_ASSETS_VALUE);
         return true;
+    }
+
+    private static boolean shouldAddSharedLibraryAssets(ApplicationInfo applicationInfo) {
+        return SDK_INT >= Build.VERSION_CODES.N && applicationInfo != null &&
+                applicationInfo.sharedLibraryFiles != null;
     }
 }
