@@ -35,6 +35,9 @@ public class ShareConstants {
     public static final String TINKER_ID     = "TINKER_ID";
     public static final String NEW_TINKER_ID = "NEW_TINKER_ID";
 
+    // Please keep it synchronized with the one defined in TypedValue.
+    public static final String PKGMETA_KEY_IS_PROTECTED_APP = "is_protected_app";
+
     public static final String OLD_VERSION     = "old";
     public static final String NEW_VERSION     = "new";
     public static final String PATCH_BASE_NAME = "patch-";
@@ -48,22 +51,23 @@ public class ShareConstants {
 
     public static final String DEX_META_FILE               = "assets/dex_meta.txt";
     public static final String DEX_PATH                    = "dex";
+    public static final String ARKHOTFIX_PATH = "arkHot";
     public static final String DEFAULT_DEX_OPTIMIZE_PATH   = "odex";
     public static final String ANDROID_O_DEX_OPTIMIZE_PATH = "oat";
 
     public static final String INTERPRET_DEX_OPTIMIZE_PATH = "interpet";
     public static final String CHANING_DEX_OPTIMIZE_PATH   = "changing";
 
-    public static final Pattern CLASS_N_PATTERN = Pattern.compile("classes(?:[2-9]{0,1}|[1-9][0-9]+)\\.dex(\\.jar)?");
-
-
     public static final String DEX_SUFFIX  = ".dex";
     public static final String JAR_SUFFIX  = ".jar";
+    public static final String APK_SUFFIX  = ".apk";
     public static final String ODEX_SUFFIX = ".odex";
 
     public static final String TEST_DEX_NAME            = "test.dex";
-    public static final String CHANGED_CLASSES_DEX_NAME = "changed_classes.dex";
     public static final String CLASS_N_APK_NAME         = "tinker_classN.apk";
+    public static final String ARKHOT_PATCH_NAME = "patch.apk";
+
+    public static final Pattern CLASS_N_PATTERN = Pattern.compile("classes(?:[2-9]?|[1-9][0-9]+)\\.dex(\\.jar)?");
 
     public static final String CHECK_DEX_INSTALL_FAIL = "checkDexInstall failed";
     public static final String CHECK_RES_INSTALL_FAIL = "checkResInstall failed";
@@ -71,7 +75,7 @@ public class ShareConstants {
     public static final String CHECK_DEX_OAT_EXIST_FAIL  = "checkDexOptExist failed";
     public static final String CHECK_DEX_OAT_FORMAT_FAIL = "checkDexOptFormat failed";
 
-//    public static final String CHECK_VM_PROPERTY_FAIL = "checkVmArtProperty failed";
+    // public static final String CHECK_VM_PROPERTY_FAIL = "checkVmArtProperty failed";
 
 
     public static final String RES_META_FILE       = "assets/res_meta.txt";
@@ -87,6 +91,7 @@ public class ShareConstants {
     public static final String RES_PATTERN_TITLE   = "pattern:";
     public static final String RES_STORE_TITLE     = "store:";
 
+    public static final String ARKHOT_META_FILE = "assets/arkHot_meta.txt";
 
     public static final String DEXMODE_RAW = "raw";
     public static final String DEXMODE_JAR = "jar";
@@ -105,14 +110,14 @@ public class ShareConstants {
      * multi process share
      */
     public static final String TINKER_SHARE_PREFERENCE_CONFIG = "tinker_share_config";
-    public static final String TINKER_ENABLE_CONFIG           = "tinker_enable_";
+    public static final String TINKER_ENABLE_CONFIG_PREFIX    = "tinker_enable_";
 
     /**
      * only for each process
      */
-    public static final String TINKER_OWN_PREFERENCE_CONFIG = "tinker_own_config_";
-    public static final String TINKER_SAFE_MODE_COUNT       = "safe_mode_count";
-    public static final int    TINKER_SAFE_MODE_MAX_COUNT   = 3;
+    public static final String TINKER_OWN_PREFERENCE_CONFIG_PREFIX = "tinker_own_config_";
+    public static final String TINKER_SAFE_MODE_COUNT_PREFIX       = "safe_mode_count_";
+    public static final int    TINKER_SAFE_MODE_MAX_COUNT          = 3;
 
 
     /**
@@ -131,14 +136,16 @@ public class ShareConstants {
     public static final int TYPE_LIBRARY     = 5;
     public static final int TYPE_RESOURCE    = 6;
     public static final int TYPE_CLASS_N_DEX = 7;
+    public static final int TYPE_ARKHOT_SO = 8;
 
 
     public static final int TINKER_DISABLE             = 0x00;
     public static final int TINKER_DEX_MASK            = 0x01;
     public static final int TINKER_NATIVE_LIBRARY_MASK = 0x02;
     public static final int TINKER_RESOURCE_MASK       = 0x04;
-    public static final int TINKER_DEX_AND_LIBRARY     = TINKER_DEX_MASK | TINKER_NATIVE_LIBRARY_MASK;
-    public static final int TINKER_ENABLE_ALL          = TINKER_DEX_MASK | TINKER_NATIVE_LIBRARY_MASK | TINKER_RESOURCE_MASK;
+    public static final int TINKER_ARKHOT_MASK = 0x08;
+    public static final int TINKER_DEX_AND_LIBRARY     = TINKER_DEX_MASK | TINKER_NATIVE_LIBRARY_MASK | TINKER_ARKHOT_MASK;
+    public static final int TINKER_ENABLE_ALL          = TINKER_DEX_MASK | TINKER_NATIVE_LIBRARY_MASK | TINKER_RESOURCE_MASK | TINKER_ARKHOT_MASK;
 
     //load error code
     public static final int ERROR_LOAD_OK                                      = 0;
@@ -169,6 +176,8 @@ public class ShareConstants {
     public static final int ERROR_LOAD_PATCH_VERSION_RESOURCE_LOAD_EXCEPTION      = -23;
     public static final int ERROR_LOAD_PATCH_VERSION_RESOURCE_MD5_MISMATCH        = -24;
     public static final int ERROR_LOAD_PATCH_UNCAUGHT_EXCEPTION                   = -25;
+    // -26 & -27 is used by WeChat internal logic.
+    public static final int ERROR_LOAD_PATCH_BAIL_HACK_FAILURE                    = -28;
 
     public static final int ERROR_LOAD_GET_INTENT_FAIL = -10000;
 
