@@ -33,23 +33,32 @@ import java.util.Properties;
 public class SharePatchInfo {
     private static final String TAG = "Tinker.PatchInfo";
 
-    public static final int    MAX_EXTRACT_ATTEMPTS = ShareConstants.MAX_EXTRACT_ATTEMPTS;
-    public static final String OLD_VERSION          = ShareConstants.OLD_VERSION;
-    public static final String NEW_VERSION          = ShareConstants.NEW_VERSION;
-    public static final String FINGER_PRINT         = "print";
-    public static final String OAT_DIR              = "dir";
-    public static final String DEFAULT_DIR   = ShareConstants.DEFAULT_DEX_OPTIMIZE_PATH;
+    public static final int    MAX_EXTRACT_ATTEMPTS        = ShareConstants.MAX_EXTRACT_ATTEMPTS;
+    public static final String OLD_VERSION                 = ShareConstants.OLD_VERSION;
+    public static final String NEW_VERSION                 = ShareConstants.NEW_VERSION;
+    public static final String IS_PROTECTED_APP            = ShareConstants.PKGMETA_KEY_IS_PROTECTED_APP;
+    public static final String IS_REMOVE_NEW_VERSION       = "is_remove_new_version";
+    public static final String FINGER_PRINT                = "print";
+    public static final String OAT_DIR                     = "dir";
+    public static final String IS_REMOVE_INTERPRET_OAT_DIR = "is_remove_interpret_oat_dir";
+    public static final String DEFAULT_DIR                 = ShareConstants.DEFAULT_DEX_OPTIMIZE_PATH;
     public String oldVersion;
     public String newVersion;
+    public boolean isProtectedApp;
+    public boolean isRemoveNewVersion;
     public String fingerPrint;
     public String oatDir;
+    public boolean isRemoveInterpretOATDir;
 
-    public SharePatchInfo(String oldVer, String newVew, String finger, String oatDir) {
+    public SharePatchInfo(String oldVer, String newVer, boolean isProtectedApp, boolean isRemoveNewVersion, String finger, String oatDir, boolean isRemoveInterpretOATDir) {
         // TODO Auto-generated constructor stub
         this.oldVersion = oldVer;
-        this.newVersion = newVew;
+        this.newVersion = newVer;
+        this.isProtectedApp = isProtectedApp;
+        this.isRemoveNewVersion = isRemoveNewVersion;
         this.fingerPrint = finger;
         this.oatDir = oatDir;
+        this.isRemoveInterpretOATDir = isRemoveInterpretOATDir;
     }
 
     public static SharePatchInfo readAndCheckPropertyWithLock(File pathInfoFile, File lockFile) {
@@ -115,8 +124,10 @@ public class SharePatchInfo {
         String oldVer = null;
         String newVer = null;
         String lastFingerPrint = null;
-        String oatDIr = null;
-
+        boolean isProtectedApp = false;
+        boolean isRemoveNewVersion = false;
+        String oatDir = null;
+        boolean isRemoveInterpretOATDir = false;
         while (numAttempts < MAX_EXTRACT_ATTEMPTS && !isReadPatchSuccessful) {
             numAttempts++;
             Properties properties = new Properties();
@@ -126,10 +137,15 @@ public class SharePatchInfo {
                 properties.load(inputStream);
                 oldVer = properties.getProperty(OLD_VERSION);
                 newVer = properties.getProperty(NEW_VERSION);
+                final String isProtectedAppStr = properties.getProperty(IS_PROTECTED_APP);
+                isProtectedApp = (isProtectedAppStr != null && !isProtectedAppStr.isEmpty() && !"0".equals(isProtectedAppStr));
+                final String isRemoveNewVersionStr = properties.getProperty(IS_REMOVE_NEW_VERSION);
+                isRemoveNewVersion = (isRemoveNewVersionStr != null && !isRemoveNewVersionStr.isEmpty() && !"0".equals(isRemoveNewVersionStr));
                 lastFingerPrint = properties.getProperty(FINGER_PRINT);
-                oatDIr = properties.getProperty(OAT_DIR);
+                oatDir = properties.getProperty(OAT_DIR);
+                final String isRemoveInterpretOATDirStr = properties.getProperty(IS_REMOVE_INTERPRET_OAT_DIR);
+                isRemoveInterpretOATDir = (isRemoveInterpretOATDirStr != null && !isRemoveInterpretOATDirStr.isEmpty() && !"0".equals(isRemoveInterpretOATDirStr));
             } catch (IOException e) {
-//                e.printStackTrace();
                 Log.w(TAG, "read property failed, e:" + e);
             } finally {
                 SharePatchFileUtil.closeQuietly(inputStream);
@@ -149,7 +165,7 @@ public class SharePatchInfo {
         }
 
         if (isReadPatchSuccessful) {
-            return new SharePatchInfo(oldVer, newVer, lastFingerPrint, oatDIr);
+            return new SharePatchInfo(oldVer, newVer, isProtectedApp, isRemoveNewVersion, lastFingerPrint, oatDir, isRemoveInterpretOATDir);
         }
 
         return null;
@@ -172,10 +188,17 @@ public class SharePatchInfo {
             + info.oldVersion
             + ", newVer:"
             + info.newVersion
+            + ", isProtectedApp:"
+            + (info.isProtectedApp ? 1 : 0)
+            + ", isRemoveNewVersion:"
+            + (info.isRemoveNewVersion ? 1 : 0)
             + ", fingerprint:"
             + info.fingerPrint
             + ", oatDir:"
-            + info.oatDir);
+            + info.oatDir
+            + ", isRemoveInterpretOATDir:"
+            + (info.isRemoveInterpretOATDir ? 1 : 0)
+        );
 
         boolean isWritePatchSuccessful = false;
         int numAttempts = 0;
@@ -191,8 +214,11 @@ public class SharePatchInfo {
             Properties newProperties = new Properties();
             newProperties.put(OLD_VERSION, info.oldVersion);
             newProperties.put(NEW_VERSION, info.newVersion);
+            newProperties.put(IS_PROTECTED_APP, (info.isProtectedApp ? "1" : "0"));
+            newProperties.put(IS_REMOVE_NEW_VERSION, (info.isRemoveNewVersion ? "1" : "0"));
             newProperties.put(FINGER_PRINT, info.fingerPrint);
             newProperties.put(OAT_DIR, info.oatDir);
+            newProperties.put(IS_REMOVE_INTERPRET_OAT_DIR, (info.isRemoveInterpretOATDir ? "1" : "0"));
 
             FileOutputStream outputStream = null;
             try {
@@ -200,7 +226,6 @@ public class SharePatchInfo {
                 String comment = "from old version:" + info.oldVersion + " to new version:" + info.newVersion;
                 newProperties.store(outputStream, comment);
             } catch (Exception e) {
-//                e.printStackTrace();
                 Log.w(TAG, "write property failed, e:" + e);
             } finally {
                 SharePatchFileUtil.closeQuietly(outputStream);
